Used bootstrap classes on html files to make them look awesome

+++ posts/views.py
@@ -5,6 +5,12 @@ from .forms import PostForm
 from django.contrib import messages
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django.contrib import messages
+from urllib.parse import quote
+from django.http import Http404
+from django.db.models import Q
+
+
+
 
 
 # Create your views here.
@@ -32,33 +38,41 @@ def post_list(request):
 	return render(request,"list.html", context)
 
 
-def post_detail(request, post_id):
-	item = Post.objects.get(id=post_id)
-	#item = get_object_or_404 (Post,id=post)
+def post_detail(request, post_slug):
+	item = Post.objects.get(slug=post_slug)
+	#item = get_object_or_404 (Post,slug=post)
 	context = {
 	"item" : item,
+	"share_string": quote(item.content)
 	}
 	return render (request, "detail.html", context)
 
 
 def post_create (request):
-	form =PostForm(request.POST or None)
+	if not request.user.is_staff:
+		raise http404
+	form = PostForm(request.POST or None, request.FILES or None)
 	if form .is_valid():
-		form.save()
+		create_post = form.save(commit=False)
+		create_post.auther = request.user
+		create_post.save()
 		messages.success (request, 'your post was added')
 		return redirect("list")
 	context = {
 		"form": form
 	}
 	return render(request, 'post_create.html' , context)
-	
-def post_update (request, post):
-	item = Post.objects.get(id=post_id)
+
+def post_update (request, post_slug):
+	if not request.user.is_staff:
+		raise Http404
+
+	item = Post.objects.get(slug=post_slug)
 	form =PostForm(request.POST or None,request.FILES or None, instance=item)
-	if form .is_valid():
+	if form.is_valid():
 		form.save()
 		messages.info(request,"you changed the post")
-		return redirect("list", post_id=item.id)
+		return redirect("list")
 	context = {
 		"form": form,
 		"item": item,
@@ -66,8 +80,11 @@ def post_update (request, post):
 	return render(request, 'post_update.html' , context)
 
 
-def post_delete(request,post_id):
-	instance = get_object_or_404(Post, id=post_id)
+def post_delete(request,post_slug):
+	if not request.user.is_staff:
+		raise Http404
+
+	instance = get_object_or_404(Post, slug=post_slug)
 	instance.delete()
 	messages.success(request, "Successfully Deleted!")
-	return redirect("Posts=list")
+	return redirect("list")
